Remove package top-level comment
Go best practices say that doc comments should start with the type they
are describing - e.g. a package comment should start with "Package xxx".

Since such a comment already exists in the "docs.go" file, I removed the
redundant ones.

If we want to emphasize to people (when browsing code) that this is the
secureworks' version of "errors", we can add a comment just below the
"package" line I suppose.